requirements.txt
python-dotenv==1.0.1
requests==2.31.0
scikit-learn>=1.4.0
apscheduler==3.10.4
flask==3.0.3
openai>=1.0.0
